First working pass of signal gen
