+++ src/applications/mne_scan/plugins/signalgen/CMakeLists.txt
@@ -11,13 +11,13 @@ find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Core Widgets Network Conc
 
 set(SOURCES
     signalgen_global.cpp
-    signalgensetupwidget.cpp
+    FormFiles/signalgensetupwidget.cpp
     signalgen.cpp
 )
 
 set(HEADERS
     signalgen_global.h
-    signalgensetupwidget.h
+    FormFiles/signalgensetupwidget.h
     signalgen.h
 )
