+++ src/applications/mne_scan/plugins/signalgen/FormFiles/signalgensetupwidget.h
